move premine-state next to similar utils
