+++ testing/util/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "deposits.go",
         "helpers.go",
         "merge.go",
+        "premine-state.go",
         "state.go",
         "sync_aggregate.go",
         "sync_committee.go",
@@ -53,6 +54,7 @@ go_library(
         "//testing/require:go_default_library",
         "//time/slots:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
+        "@com_github_ethereum_go_ethereum//core/types:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
