+++ testing/util/premine-state.go
@@ -1,4 +1,4 @@
-package spectypes
+package util
 
 import (
 	"context"
